+++ src/main/scala/com/github/simplesteph/ksm/KafkaSecurityManager.scala
@@ -19,13 +19,15 @@ object KafkaSecurityManager extends App {
   var isCancelled: AtomicBoolean = new AtomicBoolean(false)
   var grpcServer: KsmGrpcServer = _
   var aclSynchronizer: AclSynchronizer = _
+  val aclParser = new CsvAclParser(appConfig.Parser.csvDelimiter)
 
   if (appConfig.KSM.extract) {
-    new ExtractAcl(appConfig.Authorizer.authorizer, CsvAclParser).extract()
+    new ExtractAcl(appConfig.Authorizer.authorizer, aclParser).extract()
   } else {
     aclSynchronizer = new AclSynchronizer(appConfig.Authorizer.authorizer,
                                           appConfig.Source.sourceAcl,
                                           appConfig.Notification.notification,
+                                          aclParser,
                                           appConfig.KSM.readOnly)
 
     Try {