+++ src/main/scala/com/github/simplesteph/ksm/source/NoSourceAcl.scala
@@ -1,4 +1,5 @@
 package com.github.simplesteph.ksm.source
+import com.github.simplesteph.ksm.parser.AclParser
 import com.typesafe.config.Config
 
 class NoSourceAcl extends SourceAcl {
@@ -24,7 +25,7 @@ class NoSourceAcl extends SourceAcl {
     *
     * @return
     */
-  override def refresh(): Option[SourceAclResult] = None
+  override def refresh(aclParser: AclParser): Option[SourceAclResult] = None
 
   /**
     * Close all the necessary underlying objects or connections belonging to this instance