+++ src/test/scala/com/github/simplesteph/ksm/parser/CsvAclParserTest.scala
@@ -21,9 +21,10 @@ class CsvAclParserTest extends FlatSpec with Matchers {
 
   val resource = Resource(Topic, "test", PatternType.LITERAL)
   val acl = Acl(SecurityUtils.parseKafkaPrincipal("User:alice"), Allow, "*", Read)
+  val csvAclParser = new CsvAclParser(',')
 
   "parseRow" should "correctly parse a Row" in {
-    CsvAclParser.parseRow(row) shouldBe((resource, acl))
+    csvAclParser.parseRow(row) shouldBe((resource, acl))
   }
 
   "aclsFromCsv" should "correctly parse a Correct CSV" in {
@@ -44,7 +45,7 @@ class CsvAclParserTest extends FlatSpec with Matchers {
     val res2 = Resource(Group, "bar", PatternType.LITERAL)
     val res3 = Resource(Cluster, "kafka-cluster", PatternType.LITERAL)
 
-    val res = CsvAclParser.aclsFromReader(new StringReader(csv))
+    val res = csvAclParser.aclsFromReader(new StringReader(csv))
 
     res.errs shouldBe List()
 
@@ -55,7 +56,6 @@ class CsvAclParserTest extends FlatSpec with Matchers {
     )
 
   }
-
 
   "aclsFromCsv" should "catch all errors and catch all correct" in {
 
@@ -76,7 +76,7 @@ class CsvAclParserTest extends FlatSpec with Matchers {
     val res2 = Resource(Group, "bar", PatternType.LITERAL)
     val res3 = Resource(Cluster, "kafka-cluster", PatternType.LITERAL)
 
-    val res = CsvAclParser.aclsFromReader(new StringReader(csv))
+    val res = csvAclParser.aclsFromReader(new StringReader(csv))
 
     res.errs.size shouldBe 2
     val throwable1 = res.errs.head.get
@@ -103,7 +103,7 @@ class CsvAclParserTest extends FlatSpec with Matchers {
         |User:peter,Cluster,LITERAL,kafka-cluster,Create,Allow
         |""".stripMargin
 
-    val res = CsvAclParser.aclsFromReader(new StringReader(csv))
+    val res = csvAclParser.aclsFromReader(new StringReader(csv))
 
     res.errs.size shouldBe 3
 
@@ -112,7 +112,7 @@ class CsvAclParserTest extends FlatSpec with Matchers {
   "asCsv" should "correctly write CSV Row" in {
     val acl1 = Acl(SecurityUtils.parseKafkaPrincipal("User:alice"), Allow, "*", Read)
     val res1 = Resource(Topic, "foo", PatternType.LITERAL)
-    val res = CsvAclParser.asCsv(res1, acl1)
+    val res = csvAclParser.asCsv(res1, acl1)
     res shouldBe "User:alice,Topic,LITERAL,foo,Read,Allow,*"
   }
 
@@ -133,9 +133,50 @@ class CsvAclParserTest extends FlatSpec with Matchers {
     val res2 = Resource(Group, "bar", PatternType.PREFIXED)
     val res3 = Resource(Cluster, "kafka-cluster", PatternType.LITERAL)
 
-    val res = CsvAclParser.formatAcls(List((res1, acl1),(res2, acl2), (res3, acl3)))
+    val res = csvAclParser.formatAcls(List((res1, acl1),(res2, acl2), (res3, acl3)))
 
     res shouldBe csv
 
+  }
+
+
+  "aclsFromCsv" should "correctly parse a Correct CSV with different delimiter" in {
+    val funnyCsvAclParser = new CsvAclParser('?')
+    val csv =
+      """KafkaPrincipal?ResourceType?PatternType?ResourceName?Operation?PermissionType?Host
+        |User:alice?Topic?LITERAL?foo?Read?Allow?*
+        |User:bob?Group?LITERAL?bar?Write?Deny?12.34.56.78
+        |
+        |User:peter?Cluster?LITERAL?kafka-cluster?Create?Allow?*
+        |""".stripMargin
+
+
+    val acl1 = Acl(SecurityUtils.parseKafkaPrincipal("User:alice"), Allow, "*", Read)
+    val acl2 = Acl(SecurityUtils.parseKafkaPrincipal("User:bob"), Deny, "12.34.56.78", Write)
+    val acl3 = Acl(SecurityUtils.parseKafkaPrincipal("User:peter"), Allow, "*", Create)
+
+    val res1 = Resource(Topic, "foo", PatternType.LITERAL)
+    val res2 = Resource(Group, "bar", PatternType.LITERAL)
+    val res3 = Resource(Cluster, "kafka-cluster", PatternType.LITERAL)
+
+    val res = funnyCsvAclParser.aclsFromReader(new StringReader(csv))
+
+    res.errs shouldBe List()
+
+    res.acls shouldBe Set(
+      res1 -> acl1,
+      res2 -> acl2,
+      res3 -> acl3
+    )
+
+  }
+
+
+  "asCsv" should "correctly write CSV Row with different delimiter" in {
+    val funnyCsvAclParser = new CsvAclParser('?')
+    val acl1 = Acl(SecurityUtils.parseKafkaPrincipal("User:alice"), Allow, "*", Read)
+    val res1 = Resource(Topic, "foo", PatternType.LITERAL)
+    val res = funnyCsvAclParser.asCsv(res1, acl1)
+    res shouldBe "User:alice?Topic?LITERAL?foo?Read?Allow?*"
   }
 }